Refactor; split out find render areas step.
